requirements.txt
django==5.2.5
djangorestframework==3.16.1
svix
django-cors-headers==4.0.0
PyJWT==2.10.1
requests==2.32.5
cryptography==46.0.3
